Moved C++ chunk meshing code
Moved pure C++ meshing code to its own file.
