+++ src/builtin_plugins/amulet_team_3d_viewer/_view_3d/_resource_pack_base.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include <amulet/block.hpp>
 #include <amulet/mesh/block/block_mesh.hpp>
 #include <map>
